git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rebmix@255 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/rebmix/src/base.cpp
@@ -21,8 +21,8 @@ Base::~Base()
     if (length_theta_) free(length_theta_);
 } // ~Base
 
-static long IY = 0;
-static long IV[NTAB];
+static INT IY = 0;
+static INT IV[NTAB];
 
 // Minimal random number generator of Park and Miller with Bays-Durham shuffle and added
 // safeguards. Returns a uniform random deviate between 0.0 and 1.0 (exclusive of the endpoint
@@ -30,9 +30,9 @@ static long IV[NTAB];
 // between successive deviates in a sequence.RNMX should approximate the largest floating
 // value that is less than 1. See http://www.nrbook.com/a/bookcpdf/c7-1.pdf
 
-FLOAT Ran1(int *IDum)
+FLOAT Ran1(INT *IDum)
 {
-    int   j, k;
+    INT   j, k;
     FLOAT Tmp;
 
     if (*IDum <= 0 || !IY) {
@@ -63,10 +63,10 @@ FLOAT Ran1(int *IDum)
 // Inserts y into ascending list Y of length n.Set n = 0 initially.
 
 void Insert(FLOAT y,   // Inserted value.
-            int   *n,  // Length of Y.
+            INT   *n,  // Length of Y.
             FLOAT *Y)  // Pointer to Y = [y0,...,yn-1].
 {
-    int i, j;
+    INT i, j;
 
     Y[*n] = y;
 
@@ -88,7 +88,7 @@ void Insert(FLOAT y,   // Inserted value.
 FLOAT Gammaln(FLOAT y)
 {
     FLOAT x, z, Tmp, Ser;
-    int   j;
+    INT   j;
 
     static FLOAT Cof[6] = { (FLOAT)76.18009172947146, -(FLOAT)86.50532032941677,
         (FLOAT)24.01409824083091, -(FLOAT)1.231739572450155,
@@ -107,7 +107,7 @@ FLOAT Gammaln(FLOAT y)
 
 // Returns the digamma for y > 0. See http://www.nr.com/.
 
-int Digamma(FLOAT y, FLOAT *Psi)
+INT Digamma(FLOAT y, FLOAT *Psi)
 {
     static FLOAT piov4 = (FLOAT)0.785398163397448;
     static FLOAT dy0 = (FLOAT)1.461632144968362341262659542325721325;
@@ -120,11 +120,11 @@ int Digamma(FLOAT y, FLOAT *Psi)
         -(FLOAT)4.48616543918019, -(FLOAT)0.648157123766197 };
     static FLOAT q2[4] = { (FLOAT)32.2703493791143, (FLOAT)89.2920700481861,
         (FLOAT)54.6117738103215, (FLOAT)7.77788548522962 };
-    int   i, m, n, nq;
+    INT   i, m, n, nq;
     FLOAT d2;
     FLOAT w, z;
     FLOAT den, aug, sgn, ymy0, ymax, upper, ymin;
-    int Error = 0;
+    INT Error = 0;
 
     ymax = (FLOAT)INT_MAX; d2 = (FLOAT)1.0 / FLOAT_EPSILON; if (ymax > d2) ymax = d2; ymin = (FLOAT)1E-9; aug = (FLOAT)0.0;
 
@@ -147,7 +147,7 @@ int Digamma(FLOAT y, FLOAT *Psi)
                 Error = 1; goto E0;
             }
 
-            nq = (int)w; w -= (FLOAT)nq; nq = (int)(w * (FLOAT)4.0); w = (w - (FLOAT)nq * (FLOAT)0.25) * (FLOAT)4.0; n = nq / 2;
+            nq = (INT)w; w -= (FLOAT)nq; nq = (INT)(w * (FLOAT)4.0); w = (w - (FLOAT)nq * (FLOAT)0.25) * (FLOAT)4.0; n = nq / 2;
 
             if (n + n != nq) {
                 w = (FLOAT)1.0 - w;
@@ -240,14 +240,14 @@ FLOAT PoissonInv(FLOAT Fy, FLOAT Theta)
 // Returns the incomplete gamma function P(a, y) evaluated by its series
 // representation as GamSer. Also returns log(Gamma(a)) as Gamln. See http://www.nr.com/.
 
-int GammaSer(FLOAT a,       // Constant a > 0.
+INT GammaSer(FLOAT a,       // Constant a > 0.
              FLOAT y,       // Variable y > 0.
              FLOAT *GamSer, // Incomplete gamma function.
              FLOAT *Gamln)  // Log(Gamma(a)).
 {
-    int   i;
+    INT   i;
     FLOAT Sum, Del, ap;
-    int   Error = 0;
+    INT   Error = 0;
 
     *Gamln = Gammaln(a);
 
@@ -276,14 +276,14 @@ int GammaSer(FLOAT a,       // Constant a > 0.
 // Returns the incomplete gamma function Q(a, y) evaluated by its continued
 // fraction representation as GamCfg. Also returns log(Gamma(a)) as Gamln. See http://www.nr.com/.
 
-int GammaCfg(FLOAT a,       // Constant a > 0.
+INT GammaCfg(FLOAT a,       // Constant a > 0.
              FLOAT y,       // Variable y > 0.
              FLOAT *GamCfg, // Incomplete gamma function.
              FLOAT *Gamln)  // Log(Gamma(a)).
 {
-    int   i;
+    INT   i;
     FLOAT Gold, G, Fac, a0, a1, b0, b1, aif, aia, ai;
-    int   Error = 0;
+    INT   Error = 0;
 
     *Gamln = Gammaln(a);
 
@@ -323,13 +323,13 @@ int GammaCfg(FLOAT a,       // Constant a > 0.
 
 // Returns the incomplete gamma function P(a, y). Also returns log(Gamma(a)) as Gamln. See http://www.nr.com/.
 
-int GammaP(FLOAT a,       // Constant a > 0.
+INT GammaP(FLOAT a,       // Constant a > 0.
            FLOAT y,       // Variable y > 0.
            FLOAT *GamP,   // Incomplete gamma function.
            FLOAT *Gamln)  // Log(Gamma(a)).
 {
     FLOAT GamSer, GamCfg;
-    int   Error = 0;
+    INT   Error = 0;
 
     if ((y <= FLOAT_MIN) || (a <= FLOAT_MIN)) {
         *GamP = (FLOAT)0.0;
@@ -355,11 +355,11 @@ E0: return (Error);
 
 // Returns the inverse of the gamma c.d.f. for the specified Theta and Beta. See http://www.nr.com/.
 
-int GammaInv(FLOAT Fy, FLOAT Theta, FLOAT Beta, FLOAT *y)
+INT GammaInv(FLOAT Fy, FLOAT Theta, FLOAT Beta, FLOAT *y)
 {
     FLOAT dy, GamP, Gamln, Tmp;
-    int   i;
-    int   Error = 0;
+    INT   i;
+    INT   Error = 0;
 
     if (Beta > (FLOAT)1.0) {
         *y = (Beta - (FLOAT)1.0) * Theta + Eps;
@@ -424,11 +424,11 @@ FLOAT GumbelInv(FLOAT Fy, FLOAT Mean, FLOAT Sigma, FLOAT Xi)
 
 // Returns the error function erf(y). See http://www.nr.com/.
 
-int ErrorF(FLOAT y,     // Variable y.
+INT ErrorF(FLOAT y,     // Variable y.
            FLOAT *ErF)  // Error function.
 {
     FLOAT GamP, Gamln;
-    int   Error = 0;
+    INT   Error = 0;
 
     Error = GammaP((FLOAT)0.5, y * y, &GamP, &Gamln);
 
@@ -444,15 +444,15 @@ E0: return (Error);
 
 // Returns the LU decomposition of matrix A. See http://www.nr.com/
 
-int LUdcmp(int   n,     // Size of square matrix.
+INT LUdcmp(INT   n,     // Size of square matrix.
            FLOAT *A,    // Pointer to the square matrix A.
-           int   *indx, // Pointer to the permutation vector.
+           INT   *indx, // Pointer to the permutation vector.
            FLOAT *det)  // Determinant.
 {
-    int   i, imax, j, k;
+    INT   i, imax, j, k;
     FLOAT Big, Tmp;
     FLOAT *V;
-    int   Error = 0;
+    INT   Error = 0;
 
     V = (FLOAT*)malloc(n * sizeof(FLOAT));
 
@@ -517,14 +517,14 @@ E0: if (V) free(V);
 
 // Solves the set of n linear equations A x = b. See See http://www.nr.com/
 
-int LUbksb(int   n,     // Size of square matrix.
+INT LUbksb(INT   n,     // Size of square matrix.
            FLOAT *A,    // Pointer to the square matrix A.
-           int   *indx, // Pointer to the permutation vector.
+           INT   *indx, // Pointer to the permutation vector.
            FLOAT *b)    // Pointer to the solution vector.
 {
-    int   i, ii = 0, ip, j;
+    INT   i, ii = 0, ip, j;
     FLOAT Sum;
-    int   Error = 0;
+    INT   Error = 0;
 
     for (i = 0; i < n; i++) {
         ip = indx[i]; Sum = b[ip]; b[ip] = b[i];
@@ -553,16 +553,16 @@ int LUbksb(int   n,     // Size of square matrix.
 
 // Returns the determinant and the inverse matrix of A. See http://www.nr.com/
 
-int LUinvdet(int   n,     // Size of square matrix.
+INT LUinvdet(INT   n,     // Size of square matrix.
              FLOAT *A,    // Pointer to the square matrix A.
              FLOAT *Ainv, // Pointer to the inverse matrix of A.
              FLOAT *Adet) // Pointer to the determinant of A.
 {
-    int   i, *indx = NULL, j;
+    INT   i, *indx = NULL, j;
     FLOAT *b = NULL, *B = NULL;
-    int   Error = 0;
+    INT   Error = 0;
 
-    indx = (int*)calloc((size_t)n, sizeof(int));
+    indx = (INT*)calloc((size_t)n, sizeof(INT));
 
     Error = NULL == indx; if (Error) goto E0;
 
@@ -603,14 +603,14 @@ E0: if (B) free(B);
 
 // Returns the Cholesky decomposition of matrix A. See http://www.nr.com/
 
-int Choldc(int   n,   // Size of square matrix.
+INT Choldc(INT   n,   // Size of square matrix.
            FLOAT *A,  // Pointer to the square matrix A.
            FLOAT *L)  // Lower triangular factors.
 {
-    int   i, j, k;
+    INT   i, j, k;
     FLOAT Sum;
     FLOAT *p = NULL;
-    int   Error = 0;
+    INT   Error = 0;
 
     memmove(L, A, n * n * sizeof(FLOAT));
 
@@ -648,15 +648,15 @@ E0: return (Error);
 
 // Returns the determinant and the inverse matrix of A. See http ://www.nr.com/
 
-int Cholinvdet(int   n,         // Size of square matrix.
+INT Cholinvdet(INT   n,         // Size of square matrix.
                FLOAT *A,        // Pointer to the symmetric square matrix A.
                FLOAT *Ainv,     // Pointer to the inverse matrix of A.
                FLOAT *logAdet)  // Pointer to the logarithm of determinant of A.
 {
-    int   i, j, k;
+    INT   i, j, k;
     FLOAT *L = NULL, Sum;
     FLOAT *p = NULL;
-    int   Error = 0;
+    INT   Error = 0;
 
     L = (FLOAT*)malloc(n * n * sizeof(FLOAT));
 
@@ -810,7 +810,7 @@ FLOAT BesselI1(FLOAT y)
 FLOAT vonMisesCdf(FLOAT y, FLOAT Mean, FLOAT Kappa)
 {
     FLOAT A[3], Io, In, Fy;
-    int   i, Error;
+    INT   i, Error;
 
     if (y > Pi2) {
         Fy = (FLOAT)1.0;
@@ -852,7 +852,7 @@ FLOAT vonMisesCdf(FLOAT y, FLOAT Mean, FLOAT Kappa)
 FLOAT vonMisesInv(FLOAT Fy, FLOAT Mean, FLOAT Kappa)
 {
     FLOAT yl, ym, yh, fl, fm;
-    int   Stop;
+    INT   Stop;
 
     if (Fy >= (FLOAT)1.0) {
         ym = Pi2;
